src/portal/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.1
certifi==2022.5.18
charset-normalizer==2.0.12
Django==3.2.16
django-appconf==1.0.5
django-betterforms==2.0.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==23.1
django-chartjs==2.3.0
django-imagekit==4.1.0
django-mysql==3.8.1
django-session-timeout==0.1.0
django-webpack-loader==0.7.0
django-widget-tweaks==1.4.3
django-wkhtmltopdf==3.2.0
gunicorn
html5lib==1.1
idna==3.3
imgkit==1.2.2
importlib-metadata==2.1.3
mysqlclient==2.0.1
payjp==0.1.0
pdfkit==0.6.1
pilkit==2.0
Pillow==9.1.1
PyPDF2==1.26.0
pypng==0.20220715.0
PyQRCode==1.2.1
python-barcode==0.14.0
python-dateutil==2.8.1
pytz==2020.1
pyzbar==0.1.9
reportlab==3.6.9
requests==2.27.1
six==1.15.0
soupsieve==2.4
sqlparse==0.3.1
stripe==4.1.0
typing-extensions==4.5.0
urllib3==1.26.9
webencodings==0.5.1
xhtml2pdf==0.2.3